Fixed packaging by including gems and restructuring things as a module
